Rock-Paper-Scissors
This is a game which is called "Rock, Paper, Scissors" game. Well what we do in this game? Explanation: 1st - It's a game which is played between two people and they have the choise between three elements to can they play and they are: Rock, Paper, Scissor. 2nd - In this table we can see when someone wins and the possibles outcomes of the game.